+++ R/is.redundant.R
@@ -23,8 +23,43 @@
 ## confidence of their proper sub-rules (i.e., have a negative improvement).
 
 setMethod("is.redundant", signature(x = "rules"),
-  function(x, measure = "confidence") {
-    interestMeasure(x, measure = "improvement", transactions = NULL, reuse = TRUE, 
-      quality_measure = measure) <= 0
+  function(x, measure = "confidence", ...) {
+    if (measure == "oddsRatioCI")
+      .improvementCI(x, transactions = NULL, reuse = TRUE, ...) <= 0
+    else
+      interestMeasure(x, measure = "improvement", transactions = NULL, reuse = TRUE, 
+        improvementMeasure = measure, ...) <= 0
   }
 )
+
+
+### TODO: The empty LHS rule has a OR of NaN! Should it be f11/f10?
+
+# Is the supersets oddsRatio sign. larger than all its subsets?
+# I.e., the superset's lower bound needs to be larger than the subset's upper bound
+.improvementCI <- function(x,
+  transactions = NULL,
+  reuse = TRUE,
+  confidenceLevel = 0.9,
+  smoothCounts = 1) {
+
+  q <- interestMeasure(x, "oddsRatioCI", transactions, reuse, 
+    confidenceLevel = confidenceLevel, smoothCounts = smoothCounts)
+  # q is a data.frame with lowerBound and upperBound
+  
+  imp <- numeric(length(x))
+  
+  ### do it by unique rhs
+  rr <- .Call(R_pnindex, rhs(x)@data, NULL, FALSE)
+  
+  for (r in unique(rr)) {
+    pos <- which(rr == r)
+    
+    ### FALSE is for verbose
+    qsubmax <- .Call(R_pnmax, lhs(x[pos])@data, q$upperBound[pos], FALSE)
+    
+    imp[pos] <- q$lowerBound[pos] - qsubmax
+  }
+  
+  imp
+}